Fixing lm95071 negative temperature readings

+++ drivers/input/misc/lm95071.c
@@ -1,5 +1,5 @@
 /*
- * .c
+ * lm95071.c
  *
  *	Digital 14-bits temperature input/hwmon driver for LM95071 from Texas Instruments.
  *
@@ -37,7 +37,7 @@ struct lm95071_chip
 	struct input_polled_dev			*input;
 };
 
-static int lm95071_spi_read(struct lm95071_chip *this, u16 *dst)
+static int lm95071_spi_read(struct lm95071_chip *this, s16 *dst)
 {
 	int error = 0;
 
@@ -58,10 +58,10 @@ static int lm95071_spi_read(struct lm95071_chip *this, u16 *dst)
 	return 0;
 }
 
-static int lm95071_read_temperature(struct lm95071_chip *this, int *dst)
+static int lm95071_read_temperature(struct lm95071_chip *this, s32 *dst)
 {
-	s16 temp;
-	u16 raw;
+    s32 temp;
+	s16 raw;
 	int error = 0;
 
 	error = lm95071_spi_read(this, &raw);
@@ -79,7 +79,7 @@ static int lm95071_read_temperature(struct lm95071_chip *this, int *dst)
 	 *
 	 * Official docs here: http://www.ti.com/lit/ds/symlink/lm95071.pdf
 	 */
-	temp = (s16)(raw / 4);
+	temp = (s32)(raw / 4);
 	temp *= 31;
 
 	dev_dbg(&(this->spi->dev), "temp read: %d mC\n", temp);
